Replace Avatar.Feedback with more generic Feedback component
