+++ src/components/Avatar/Avatar.tsx
@@ -2,11 +2,10 @@ import React from 'react'
 import { Howl } from 'howler'
 import { motion } from 'framer-motion'
 import { rng } from 'src/functions'
-import { Wrapper, PortraitImg, Name, Flash, FeedbackText } from './AvatarStyles'
+import { Wrapper, PortraitImg, Name, Flash } from './AvatarStyles'
 import impactSound from 'src/sounds/impact.slice.wav'
 
 const DAMAGE_FLASH_DURATION = 0.33
-const DAMAGE_FEEDBACK_DURATION = 0.8
 
 interface AvatarProps {
   children: any
@@ -19,10 +18,6 @@ interface PortraitProps {
 
 interface PortraitWrapperProps {
   isTakingDamage: boolean
-  children: any
-}
-
-interface FeedbackProps {
   children: any
 }
 
@@ -90,35 +85,11 @@ function DamageFlash() {
   )
 }
 
-function Feedback(props: FeedbackProps) {
-  const { children } = props
-
-  return (
-    <motion.div
-      style={{
-        position: 'absolute',
-        top: '25%',
-        right: '25%',
-      }}
-      animate={{
-        y: [0, -50, -100],
-        scale: [1, 1.25, 1.5],
-        opacity: [0, 1, 0],
-      }}
-      transition={{ duration: DAMAGE_FEEDBACK_DURATION }}
-    >
-      <FeedbackText>-{children}</FeedbackText>
-    </motion.div>
-  )
-}
-
 Portrait.displayName = 'Avatar.Portrait'
 Avatar.Portrait = Portrait
 Name.displayName = 'Avatar.Name'
 Avatar.Name = Name
 Flash.displayName = 'Avatar.Flash'
 Avatar.Flash = Flash
-Feedback.displayName = 'Avatar.Feedback'
-Avatar.Feedback = Feedback
 
 export default Avatar
