+++ src/components/Avatar/AvatarStyles.ts
@@ -40,12 +40,3 @@ export const Flash = styled(motion.div)`
   height: 100%;
   background-color: ${props => props.theme.colors.white};
 `
-
-export const FeedbackText = styled.div`
-  font-family: ${props => props.theme.fonts.heading};
-  font-weight: 700;
-  font-size: ${props => props.theme.fontSizes[3]};
-  color: ${props => props.theme.colors.pink};
-  text-shadow: 0 0 15px ${props => props.theme.colors.white},
-    0 2px 0 ${props => props.theme.colors.darkPink};
-`
