+++ src/components/Feedback/index.ts
@@ -1,0 +1,1 @@
+export { Feedback } from './Feedback'
