+++ src/components/index.ts
@@ -5,6 +5,7 @@ export * from './Button'
 export * from './Card'
 export * from './CardRadioButton'
 export * from './Deck'
+export * from './Feedback'
 export * from './Game'
 export * from './Label'
 export * from './Legend'
